Спринт 16. Начинаю переделывать Search. Распределил старые файлы по пакетам
