--- app/src/main/java/com/example/playlistmaker/data/old/NetworkClient.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.data.old
-
-interface NetworkClient {
-    fun doRequest(dto: Any): Response
-}
