--- app/src/main/java/com/example/playlistmaker/data/old/Responce.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.data.old
-
-open class Response() {
-    var resultCode = 0
-}
