--- app/src/main/java/com/example/playlistmaker/data/settings/impl/SettingsRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.data.settings.impl
-
-class SettingsRepositoryImpl {
-}
