--- app/src/main/java/com/example/playlistmaker/data/sharing/ExternalNavigator.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.data.sharing
-
-interface ExternalNavigator {
-}
