--- app/src/main/java/com/example/playlistmaker/data/sharing/impl/ExternalNavigatorImpl.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.data.sharing.impl
-
-class ExternalNavigatorImpl {
-}
