--- app/src/main/java/com/example/playlistmaker/domain_old/old/TracksRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain_old.old
-
-import com.example.playlistmaker.search.ui.model.Track
-
-interface TracksRepository {
-    fun searchTrack(expression: String): List<Track>
-}
