--- app/src/main/java/com/example/playlistmaker/domain_old/settings/SettingsInteractor.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain_old.settings
-
-import com.example.playlistmaker.domain_old.settings.model.ThemeSettings
-
-interface SettingsInteractor {
-    fun getThemeSettings(): ThemeSettings
-    fun updateThemeSetting(settings: ThemeSettings)
-}
