--- app/src/main/java/com/example/playlistmaker/domain_old/settings/model/ThemeSettings.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.domain_old.settings.model
-
-class ThemeSettings {
-}
