--- app/src/main/java/com/example/playlistmaker/old/TrackView.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.old
-
-interface TrackView {
-}
