+++ app/src/main/java/com/example/playlistmaker/player/domain/impl/PlayerInteractorImpl.kt
@@ -1,6 +1,5 @@
 package com.example.playlistmaker.player.domain.impl
 
-import com.example.playlistmaker.domain_old.old.TracksRepository
 import com.example.playlistmaker.player.domain.api.PlayerInteractor
 import com.example.playlistmaker.player.domain.api.PlayerRepository
