+++ app/src/main/java/com/example/playlistmaker/player/ui/activity/PlayerActivity.kt
@@ -1,12 +1,6 @@
 package com.example.playlistmaker.player.ui.activity
 
-import android.media.MediaPlayer
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import android.util.Log
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.bumptech.glide.Glide
