+++ app/src/main/java/com/example/playlistmaker/search/SearchTrackHistory.kt
@@ -1,8 +1,8 @@
-package com.example.playlistmaker.ui.old
+package com.example.playlistmaker.search
 
 import android.content.SharedPreferences
 import android.util.Log
-import com.example.playlistmaker.search.ui.model.Track
+import com.example.playlistmaker.search.domain.model.Track
 import com.google.gson.Gson
 
 const val HISTORY_PREFS = "Playlist_Maker_History"
