+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TrackDto.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.search.data.dto
+
+data class TrackDto(
+    val trackId: String,
+    val trackName: String,
+    val artistName: String,
+    val trackTimeMillis: String,
+    val artworkUrl100: String,
+    val collectionName: String,
+    val releaseDate: String,
+    val primaryGenreName: String,
+    val country: String,
+    val previewUrl: String
+)
