+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TrackSearchResponse.kt
@@ -1,3 +1,3 @@
-package com.example.playlistmaker.data.old
+package com.example.playlistmaker.search.data.dto
 
 class TrackSearchResponse(val results: List<TrackDto>) : Response()
