+++ app/src/main/java/com/example/playlistmaker/search/data/network/ITunesAPI.kt
@@ -1,5 +1,6 @@
-package com.example.playlistmaker.data.old
+package com.example.playlistmaker.search.data.network
 
+import com.example.playlistmaker.search.data.dto.TrackSearchResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
