+++ app/src/main/java/com/example/playlistmaker/search/data/storage/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.search.data.storage
+
+import com.example.playlistmaker.search.data.dto.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+}
