+++ app/src/main/java/com/example/playlistmaker/search/domain/impl/TracksInteractorImpl.kt
@@ -1,5 +1,7 @@
-package com.example.playlistmaker.domain_old.old
+package com.example.playlistmaker.search.domain.impl
 
+import com.example.playlistmaker.search.domain.api.TracksInteractor
+import com.example.playlistmaker.search.domain.api.TracksRepository
 import java.util.concurrent.Executors
 
 class TracksInteractorImpl(private val repository: TracksRepository): TracksInteractor {
