+++ app/src/main/java/com/example/playlistmaker/search/ui/model/TrackUi.kt
@@ -3,7 +3,7 @@ package com.example.playlistmaker.search.ui.model
 import java.text.SimpleDateFormat
 import java.util.Locale
 
-data class Track(
+data class TrackUi(
     val trackId: String,
     val trackName: String,
     val artistName: String,
